Add threading for remove log cut
Add threading for remove log cut to speed up downloading and cutting the remove files. The same should be done for the local files in the future.
Also, separate downloading remote files via rsync, to avoid issue with downloading big files. It doesn't solve the problem completely, but it's a decent mitigation for now.